Porting to struts 2 2.3.x and Hibernate 5.2.x
Several change to make things build and run properly with updated
dependencies.

+++ voms-admin-server/src/main/java/org/glite/security/voms/admin/persistence/dao/hibernate/AuditSearchDAOHibernate.java
@@ -138,10 +138,10 @@ public class AuditSearchDAOHibernate
     crit.setFirstResult(0);
     crit.setMaxResults(100);
 
-    Integer count = (Integer) crit.uniqueResult();
+    Long count = (Long) crit.uniqueResult();
 
     if (count == null) {
-      count = 0;
+      count = 0L;
     }
 
     return new AuditLogSearchResults(sp, count, results);
